+++ Locales/enUS.lua
@@ -72,10 +72,20 @@ select(2, ...).L = setmetatable({
     ["SNIPPETS_TIPS"] = "[Double-Click] to rename. [Shift-Click] to delete. All checked snippets will be automatically invoked at the end of Cell initialization process (in ADDON_LOADED event).",
 
     ["CHANGELOGS"] = [[
+        <h1>Try disabling code snippets, if Cell doesn't work after update.</h1>
+
         <h1>About the M+ Afflicted Souls</h1>
         <p>I've received some requests about showing Afflicted Souls on Cell. Simply put, due to the limitation of the plugin API, it is not possible. I can make them display on Cell, but these buttons will not be clickable, so there is no need. It is better to use WA.</p>
         <br/>
 
+        <h1>r174-release (Jun 18, 2023, 17:25 GMT+8)</h1>
+        <p>* Updated font options for some indicators. The stack font and duration font can be set separately. And if you use CooldownIcons_BlizzardStyle snippet, an update is required.</p>
+        <p>* Health Text indicator: added options to show shield value.</p>
+        <p>* Dispels indicator: updated Highlight Type option.</p>
+        <p>* Fixed Private Auras indicator.</p>
+        <p>* Updated raid debuffs.</p>
+        <br/>
+        
         <h1>r173-release (Jun 2, 2023, 18:36 GMT+8)</h1>
         <p>* Added a "Hide in combat" option for Leader Icon indicator.</p>
         <p>* Updated zhTW.</p>