+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r172-release...764ad4a8f4698dd3f8858b33396369b5c472beda)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r173-release...ac89c467936e23fc93268353fdf92a25696e34c7)
 
-- Add a hideInCombat option for Leader Icon
-- Update zhTW
+- Update raid debuffs
+- Add options to show shield value with Health Text indicator
+- Update font options for icon indicators
+- Add gradient-half for Dispels
+- Fix auraIndex of privateAuras
+- Update font settings for CooldownIcons_BlizzardStyle snippet
+- Update click-casting for Soul Stone
+- Increase scroll step for indicator settings